Move container/auto blocks into modular_blocks
